+++ docs/changelog.md
@@ -6,6 +6,7 @@
 
 - [pre-commit](https://pre-commit.com/) utility usage
 - `datastructures.LinkHeader.rel` property
+- indexed parameter lookup in `datastructures.LinkHeader`
 
 ### Changed